+++ packages/twenty-server/src/core/api-rest/api-rest-query-builder/factories/input-factories/filter-utils/add-default-conjunction.utils.ts
@@ -1,6 +1,10 @@
+import { Conjunctions } from 'src/core/api-rest/api-rest-query-builder/factories/input-factories/filter-utils/parse-filter.utils';
+
+export const DEFAULT_CONJUNCTION = Conjunctions.and;
+
 export const addDefaultConjunctionIfMissing = (filterQuery: string): string => {
   if (!(filterQuery.includes('(') && filterQuery.includes(')'))) {
-    return `and(${filterQuery})`;
+    return `${DEFAULT_CONJUNCTION}(${filterQuery})`;
   }
 
   return filterQuery;
