+++ packages/twenty-server/src/core/open-api/utils/get-error-responses.utils.ts
@@ -1,0 +1,15 @@
+export const getErrorResponses = (description: string) => {
+  return {
+    description,
+    content: {
+      'application/json': {
+        schema: {
+          type: 'object',
+          properties: {
+            error: { type: 'string' },
+          },
+        },
+      },
+    },
+  };
+};
